refactor!: split gks-common module into domain + entity
* ga4gh.core.models split into ga4gh.core.domain_models and ga4gh.core.entity_models
